requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.100.0
psycopg[binary]>=3.1.0
PyYAML>=6.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
